Add support for absolute file:// URIs and add two simple test cases. Not
* src/uri.c: (gnet_uri_new), (gnet_uri_get_string):
* tests/uri_test.c: (main):
Add support for absolute file:// URIs and add two simple test cases.
Not entirely convinced about the required fix to _get_string(), but
it doesn't seem to break any of the tests, so it can't be too bad
even if it's not entirely correct. Fixes bug #317456. Based
on patch by Milosz Derezynski <internalerror at gmail com>